chore: Factor out initial state handling from Problem.
